+++ src/com/qozix/tileview/TileView.java
@@ -31,8 +31,8 @@ import com.qozix.tileview.paths.PathManager;
 import com.qozix.tileview.samples.SampleManager;
 import com.qozix.tileview.tiles.TileManager;
 import com.qozix.tileview.tiles.TileRenderListener;
-import com.qozix.tileview.tiles.selection.ITileSelection;
-import com.qozix.tileview.tiles.selection.TileSelectionNoScale;
+import com.qozix.tileview.tiles.selector.TileSetSelector;
+import com.qozix.tileview.tiles.selector.TileSetSelectorMinimalUpScale;
 
 /**
  * The TileView widget is a subclass of ViewGroup that provides a mechanism to asynchronously display tile-based images,
@@ -1149,16 +1149,16 @@ public class TileView extends ZoomPanLayout {
 		
 	}
 
-	public ITileSelection getTileSelection() {
-	    return this.detailManager.getTileSelection();
+	public TileSetSelector getTileSetSelector() {
+	    return this.detailManager.getTileSetSelector();
 	}
 
 	/**
-	 * Set the tile selection method, defaults to {@link TileSelectionNoScale}
+	 * Set the tile selection method, defaults to {@link TileSetSelectorMinimalUpScale}
 	 * 
 	 * @param selector
 	 */
-	public void setTileSelection(ITileSelection selector) {
-	    this.detailManager.setTileSelection(selector);
+	public void setTileSetSelector(TileSetSelector selector) {
+	    this.detailManager.setTileSetSelector(selector);
 	}
 }
