+++ src/com/qozix/tileview/tiles/selector/TileSetSelector.java
@@ -1,8 +1,8 @@
-package com.qozix.tileview.tiles.selection;
+package com.qozix.tileview.tiles.selector;
 
 import com.qozix.tileview.detail.DetailLevel;
 import com.qozix.tileview.detail.DetailLevelSet;
 
-public interface ITileSelection {
+public interface TileSetSelector {
     public DetailLevel find(double scale, DetailLevelSet levels);
 }
